// src/app/performences/models/performence-pie-model.ts
export interface PerformencePieModel {
    ponctualite: number;
    hospitalite: number;
    travail: number;
    month: number;
}

export interface PerformencePieYearModel {
    ponctualite: number;
    hospitalite: number;
    travail: number;
    year: number;
}

export interface PerformencePieDayModel {
    ponctualite: number;
    hospitalite: number;
    travail: number;
    day: number;
}